Refactor user app
