+++ spotseeker/user/serializers/email.py
@@ -1,0 +1,5 @@
+from rest_framework import serializers
+
+
+class EmailSerializer(serializers.Serializer):
+    email = serializers.EmailField()
