+++ spotseeker/user/serializers/user.py
@@ -1,14 +1,9 @@
 from rest_framework import serializers
 
-from spotseeker.user.models import Notification
 from spotseeker.user.models import User
-from spotseeker.user.models import UserOTP
 
 
 class UserSerializer(serializers.ModelSerializer):
-    followers = serializers.IntegerField(source="followers.count()", read_only=True)
-    following = serializers.IntegerField(source="following.count()", read_only=True)
-
     class Meta:
         model = User
         fields = [
@@ -24,8 +19,6 @@ class UserSerializer(serializers.ModelSerializer):
             "created_at",
             "updated_at",
             "deleted_at",
-            "followers",
-            "following",
         ]
 
         extra_kwargs = {
@@ -35,12 +28,6 @@ class UserSerializer(serializers.ModelSerializer):
             "deleted_at": {"read_only": True},
             "is_validated": {"read_only": True},
         }
-
-
-class UserOTPSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = UserOTP
-        fields = ["otp"]
 
 
 class UserPasswordUpdateSerializer(serializers.ModelSerializer):
@@ -55,13 +42,3 @@ class RecoverPasswordSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ["password"]
-
-
-class NotificationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Notification
-        fields = ["user", "user_interaction", "content"]
-
-
-class EmailSerializer(serializers.Serializer):
-    email = serializers.EmailField()
