+++ spotseeker/user/views/__init__.py
@@ -1,0 +1,14 @@
+from .create import UserCreateView
+from .notification import NotificationView
+from .password import RecoverPasswordOTPView
+from .password import RecoverPasswordView
+from .user import UserViewSet
+
+__all__ = [
+    "UserCreateView",
+    "UserViewSet",
+    "UserPasswordUpdateView",
+    "RecoverPasswordView",
+    "RecoverPasswordOTPView",
+    "NotificationView",
+]
