+++ spotseeker/user/views/notification.py
@@ -1,0 +1,17 @@
+from rest_framework.mixins import ListModelMixin
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.viewsets import GenericViewSet
+from rest_framework_simplejwt.authentication import JWTAuthentication
+
+from spotseeker.user.models import Notification
+from spotseeker.user.serializers import NotificationSerializer
+
+
+class NotificationView(GenericViewSet, ListModelMixin):
+    authentication_classes = [JWTAuthentication]
+    permission_classes = [IsAuthenticated]
+    serializer_class = NotificationSerializer
+    queryset = Notification.objects.all()
+
+    def get_queryset(self, *args, **kwargs):
+        return self.queryset.filter(user=self.request.user)
